Game StateMachine + Game Manager + Cube Generation +Game engine
Developped several functions for the game to run.

Default phases have been set up.
Moved the cube generation into a different script that is called after
the informations are added to the arrays.

Small design changes for the tiles.

Fixed the cube duplication code. They are generated once.

preparing the GameEngine so that it moves the blocks down  a line every
1 second.

- Need to add a TAG to each cube and attach the tags upon creation.