Checking each row
Made a loop that check every row, ( and the rows only ). It
automatically stops when it reaches the end of the table.

Will need to add that it checks if the position below is available or
not. If so it will go lower. if not, it will stop the iteration.