Game over when table filled
Check the number of availlable spots before generating the cubes. if
the is not enough space to generate the cubes we show a game over
screen.

For now the screen is temporary. Will need to prevent the cubes from
moving and put a 1 second delay than move to the real game over screen
with the help of the FSM.