Combi of 3 or 4
Fixed the issue where only combi of 3 or 4 were detected. Now it does
detect bothe with no issue. Need to test to see if it works with other
cube colors/Cube sets. Need to think about the gameplay whether cubes
of only one direction should be deleted or all directions.  Will start
preparing the game to input the up and down buttons in order to change
the rows cube order.