Movement Up
Worked on the cube up/down movement, it is now moving up, or seems to
be doing to do so. Need to call the refresh so function that it is
refresh upon each mouse click. Need to add the code for the buttom so
that it does work when clicking on the bottom button.