Testing on actual hardware
Game is now moving. Gave up on the multidirectional detection. Fixed
the issue where the field was not showing up upon reset. Added an extra
track for the credits changed the name of the game from Project Pomme
to CHUKU CHUKU (the sounds of the bells in the music) might change
thename for something cutier.